# services-java/mcp-tool-server/src/main/resources/openapi/mock-provider-api.yaml
openapi: 3.0.3
info:
  title: Mock Commerce Provider API
  description: |
    Mock Provider API for testing and development of the Commerce AI platform.

    This provider simulates a real e-commerce platform with a comprehensive product catalog,
    cart management, order processing, reviews, and recommendations.

    ## Features
    - **40+ Product Catalog** across Electronics, Fashion, and other categories
    - **Full Shopping Experience**: Search, Cart, Checkout, Orders
    - **Product Reviews & Ratings** with mock data
    - **Recommendations Engine** (similar, complementary, trending products)
    - **Promotions & Coupons** with validation
    - **Shipping Estimates** based on location
    - **Product Variants** (color, size, storage)

    ## Mock Data Characteristics
    - Network delay simulation (150ms per request)
    - Realistic product attributes and prices (INR)
    - In-memory storage (resets on restart)
    - Deterministic behavior for testing

  version: 1.0.0
  contact:
    name: Commerce AI Team
    email: commerce-ai@acme.com
servers:
  - url: http://localhost:8081/api/v1/mock
    description: Local development (via MCP Tool Server)
...
paths:
...
  /cart/{userId}/items:
    post:
      tags:
        - Cart
      summary: Add item to cart
      description: |
        Add a product to the user's shopping cart.

        **Behavior:**
        - If item already exists, quantity is incremented
        - Cart totals are automatically recalculated
        - Product availability is validated
      operationId: mockAddToCart
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - quantity
              properties:
                productId:
                  type: string
                  example: "MOB001"
                quantity:
                  type: integer
                  minimum: 1
                  example: 1
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Product not found
  /cart/{userId}/items/{productId}:
    put:
      tags:
        - Cart
      summary: Update cart item quantity
      description: Update the quantity of an item in the cart (set to 0 to remove)
      operationId: mockUpdateCartItem
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'

    delete:
      tags:
        - Cart
      summary: Remove item from cart
      description: Remove a product from the shopping cart
      operationId: mockRemoveCartItem
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
...
components:
  schemas:
    ProductSummary:
      type: object
      required:
        - id
        - provider
        - name
        - price
        - availability
      properties:
        id:
          type: string
          example: "MOB001"
        provider:
          type: string
          enum: [mock]
          example: "mock"
        name:
          type: string
          example: "Vivo T2x 5G 128 GB, 8 GB RAM, Glimmer Black"
        description:
          type: string
        brand:
          type: string
          example: "VIVO"
        category:
          type: string
          example: "electronics.mobile.smartphones"
        price:
          $ref: '#/components/schemas/Money'
        imageUrl:
          type: string
          format: uri
          example: "https://cdn.example.com/products/MOB001.jpg"
        availability:
          $ref: '#/components/schemas/Availability'
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.3
        reviewCount:
          type: integer
          example: 324
        attributes:
          type: object
          additionalProperties: true
          description: Category-specific attributes
          example:
            storage: "128 GB"
            ram: "8 GB"
            screenSize: "6.58 inch"
            warranty: "1 Year"
            tags: "5G,BESTSELLER"
...
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: decimal
          example: 12749
        currency:
          type: string
          example: "INR"
    Availability:
      type: object
      required:
        - inStock
        - status
      properties:
        inStock:
          type: boolean
        quantity:
          type: integer
          description: Available quantity (null if not disclosed)
        status:
          type: string
          enum: [IN_STOCK, OUT_OF_STOCK, LOW_STOCK, PREORDER]
...
    Cart:
      type: object
      required:
        - id
        - userId
        - items
        - total
      properties:
        id:
          type: string
          example: "CART-user123"
        userId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        currency:
          type: string
          example: "INR"
        itemCount:
          type: integer
          example: 2
    CartItem:
      type: object
      required:
        - productId
        - provider
        - quantity
        - unitPrice
      properties:
        productId:
          type: string
        provider:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          $ref: '#/components/schemas/Money'
        product:
          $ref: '#/components/schemas/ProductSummary'
...